+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Fix: `TextButton`: Improve styles.
 - Fix: `DashboardLayout`: Fix main wrapper focus style.
 - Fix: `DocumentsManager`: Fix markup.
+- Feature: `BackingCard`: Add `truncateText` prop to `BackingCard.Title`.
 - Fix: Markup: Fix duplicate ID issue on the following SVG components:
   - `GoodeedLogo`
   - `PhoneIllustration`